chore(codegen): #1098 split expr.rs — extract channel + i32_fast_path submodules (behavior-preserving)
